refactor: removed mock data, fixed types
