+++ common/client/fragments/header-fragment/src/data/index.ts
@@ -1,3 +1,4 @@
+export * from './data.interfaces'
 export * from './decide-project-team-join-request'
 export * from './decide-team-join-request'
 export * from './decide-user-team-join-invite'
