+++ common/client/fragments/header-fragment/src/notifications/notification-list/notification-item/notification-item.constants.ts
@@ -1,0 +1,7 @@
+import { NotificationTypes } from '../../../data'
+
+export const decideTypes = [
+  NotificationTypes.TeamUserJoinRequest,
+  NotificationTypes.TeamJoinProjectRequest,
+  NotificationTypes.TeamInvite,
+]
